dockerTools: Propagate image contents via passthru when using Nix DB
This commit allows for better composition of image layers which contain
Nix databases. Derivations produced by buildLayeredImageWithNixDb will
have a list of its contents appended to passthru.propagatedContents.
If the resulting derivation is passed to buildLayeredImageWithNixDb
via fromImage, the Nix database will be populated with the store paths
from all layers, not just the deepest layer.

This is useful because it allows one to extend the `nixos/nix` Docker
Hub image with additional packages and have them appear properly in
the output of `nix profile list`.

Example:

> nixpkgs = builtins.getFlake "nixpkgs"
> pkgs = import nixpkgs {}
> x = pkgs.dockerTools.buildLayeredImageWithNixDb {
    name = "foo";
    tag = "latest";
    contents = [ pkgs.bashInteractive ];
    maxLayers = 2;
}
> y = pkgs.dockerTools.buildLayeredImageWithNixDb {
    fromImage = x;
    name = "bar";
    tag = "latest";
    contents = [ pkgs.coreutils ];
    maxLayers = 4;
}
> y.propagatedContents
[ «derivation ...coreutils-9.1.drv» «derivation ...bash-interactive-5.1-p16.drv» ]